+++ frontend/src/components/google-maps/map/CarFfeineMap.tsx
@@ -1,6 +1,23 @@
+import { useEffect } from 'react';
+
+import { useQueryClient } from '@tanstack/react-query';
+
 import StationMarkersContainer from '@marker/StationMarkersContainer';
 
+import { debounce } from '@utils/debounce';
+import { useExternalValue, useSetExternalState } from '@utils/external-state';
+import { setLocalStorage } from '@utils/storage';
+
+import { getGoogleMapStore } from '@stores/google-maps/googleMapStore';
+import { warningModalActions } from '@stores/layout/warningModalStore';
+import { popupMenuOpenStore } from '@stores/popupMenuOpenStore';
+import { serverStationFilterAction } from '@stores/station-filters/serverStationFiltersStore';
+
+import { useCarFilters } from '@hooks/tanstack-query/station-filters/useCarFilters';
+import { useMemberFilters } from '@hooks/tanstack-query/station-filters/useMemberFilters';
+
 import ToastContainer from '@common/Toast/ToastContainer';
+import ZoomWarningModal from '@common/WarningModal';
 
 import ClientStationFilters from '@ui/ClientStationFilters';
 import MapController from '@ui/MapController';
@@ -8,8 +25,9 @@ import ModalContainer from '@ui/ModalContainer';
 import Navigator from '@ui/Navigator';
 import WarningModalContainer from '@ui/WarningModalContainer';
 
-import CarFfeineMapListener from './CarFfeineListener';
-import UserFilterListener from './UserFilterListener';
+import { INITIAL_ZOOM_SIZE } from '@constants/googleMaps';
+import { QUERY_KEY_STATION_MARKERS } from '@constants/queryKeys';
+import { LOCAL_KEY_LAST_POSITION } from '@constants/storageKeys';
 
 const CarFfeineMap = () => {
   return (
@@ -28,6 +46,55 @@ const CarFfeineMap = () => {
       <StationMarkersContainer />
     </>
   );
+};
+
+const CarFfeineMapListener = () => {
+  const googleMap = useExternalValue(getGoogleMapStore());
+  const queryClient = useQueryClient();
+  const setIsPopupMenuOpen = useSetExternalState(popupMenuOpenStore);
+
+  const debouncedIdleHandler = debounce(() => {
+    if (googleMap.getZoom() < INITIAL_ZOOM_SIZE) {
+      warningModalActions.openModal(<ZoomWarningModal />);
+    } else {
+      warningModalActions.closeModal();
+    }
+
+    queryClient.invalidateQueries({ queryKey: [QUERY_KEY_STATION_MARKERS] });
+    setIsPopupMenuOpen(false);
+
+    setLocalStorage<google.maps.LatLngLiteral>(LOCAL_KEY_LAST_POSITION, {
+      lat: googleMap.getCenter().lat(),
+      lng: googleMap.getCenter().lng(),
+    });
+  }, 300);
+
+  useEffect(() => {
+    googleMap.addListener('idle', debouncedIdleHandler);
+
+    const initMarkersEvent = googleMap.addListener('bounds_changed', async () => {
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEY_STATION_MARKERS] });
+
+      google.maps.event.removeListener(initMarkersEvent);
+    });
+  }, []);
+
+  return <></>;
+};
+
+const UserFilterListener = () => {
+  const queryClient = useQueryClient();
+  const { data: memberFilters } = useMemberFilters();
+  const { data: carFilters } = useCarFilters();
+  const { setAllServerStationFilters } = serverStationFilterAction;
+
+  if (memberFilters !== undefined) {
+    setAllServerStationFilters(memberFilters);
+    setAllServerStationFilters(carFilters);
+    queryClient.invalidateQueries([{ queryKey: [QUERY_KEY_STATION_MARKERS] }]);
+  }
+
+  return <></>;
 };
 
 export default CarFfeineMap;
