+++ frontend/src/constants/googleMaps.ts
@@ -2,5 +2,4 @@ export const DEFAULT_CENTER = {
   lat: 37.5056102333107,
   lng: 127.05081496722168,
 } as const;
-export const INITIAL_ZOOM_SIZE = 16;
-export const DELTA_FACTOR = 2;
+export const INITIAL_ZOOM_SIZE = 17;
