+++ frontend/src/constants/storageKeys.ts
@@ -6,4 +6,3 @@ export const LOCAL_KEY_TOKEN = 'CARFFEINE_TOKEN';
 export const SESSION_KEY_REPORTED_STATIONS = 'CARFFEINE_REPORTED_STATIONS';
 export const SESSION_KEY_MEMBER_TOKEN = 'CARFFEINE_MEMBER_TOKEN';
 export const SESSION_KEY_MEMBER_INFO = 'CARFFEINE_MEMBER_INFO';
-export const SESSION_KEY_LAST_REQUEST_POSITION = 'CARFFEINE_LAST_REQUEST_POSITION';
