+++ frontend/src/hooks/tanstack-query/station-markers/useStationMarkers.ts
@@ -6,7 +6,6 @@ import { getTypedObjectFromEntries } from '@utils/getTypedObjectFromEntries';
 import { getTypedObjectKeys } from '@utils/getTypedObjectKeys';
 import { getDisplayPosition } from '@utils/google-maps';
 import { getQueryFormattedUrl } from '@utils/request-query-params';
-import { setSessionStorage } from '@utils/storage';
 
 import { serverUrlStore } from '@stores/config/serverUrlStore';
 import { getGoogleMapStore } from '@stores/google-maps/googleMapStore';
@@ -19,42 +18,26 @@ import {
 
 import { DELIMITER } from '@constants';
 import { COMPANIES } from '@constants/chargers';
-import { DELTA_FACTOR, INITIAL_ZOOM_SIZE } from '@constants/googleMaps';
+import { INITIAL_ZOOM_SIZE } from '@constants/googleMaps';
 import { QUERY_KEY_STATION_MARKERS } from '@constants/queryKeys';
-import { SESSION_KEY_LAST_REQUEST_POSITION } from '@constants/storageKeys';
 
-import type { DisplayPosition } from '@type';
-import type { StationMarker } from '@type/stations';
-
-const isMapLoaded = (displayPosition: DisplayPosition) => {
-  const { latitudeDelta, longitudeDelta } = displayPosition;
-
-  return !(latitudeDelta === 0 && longitudeDelta === 0);
-};
+import type { DisplayPosition, StationMarker } from '@type/stations';
 
 export const fetchStationMarkers = async () => {
   const googleMap = getStoreSnapshot(getGoogleMapStore());
   const displayPosition = getDisplayPosition(googleMap);
+  const { latitudeDelta, longitudeDelta } = displayPosition;
 
-  if (!isMapLoaded(displayPosition)) {
+  if (latitudeDelta === 0 && longitudeDelta === 0) {
     throw new Error('지도가 로드되지 않았습니다');
   }
 
-  const requestPositionParams: DisplayPosition = {
-    ...displayPosition,
-    latitudeDelta: displayPosition.latitudeDelta * DELTA_FACTOR,
-    longitudeDelta: displayPosition.longitudeDelta * DELTA_FACTOR,
-  };
-
-  if (requestPositionParams.zoom < INITIAL_ZOOM_SIZE) {
-    setSessionStorage<DisplayPosition>(SESSION_KEY_LAST_REQUEST_POSITION, null);
-
+  if (displayPosition.zoom < INITIAL_ZOOM_SIZE) {
     return new Promise<StationMarker[]>((resolve) => resolve([]));
   }
 
-  const displayPositionKeys = getTypedObjectKeys<DisplayPosition>(requestPositionParams);
-  const displayPositionValues = Object.values(requestPositionParams).map(String);
-
+  const displayPositionKeys = getTypedObjectKeys<DisplayPosition>(displayPosition);
+  const displayPositionValues = Object.values(displayPosition).map(String);
   const displayPositionString = getTypedObjectFromEntries(
     displayPositionKeys,
     displayPositionValues
@@ -78,8 +61,6 @@ export const fetchStationMarkers = async () => {
   const stationMarkers = await fetch(`${serverUrl}/stations?${requestQueryParams}`, {
     method: 'GET',
   }).then<StationMarker[]>(async (response) => {
-    setSessionStorage<DisplayPosition>(SESSION_KEY_LAST_REQUEST_POSITION, requestPositionParams);
-
     const data = await response.json();
 
     return data.stations;
