+++ frontend/src/types/index.ts
@@ -2,4 +2,3 @@ export * from './stations';
 export * from './chargers';
 export * from './congestion';
 export * from './style';
-export { DisplayPosition } from '@type/map';
