+++ frontend/src/utils/getTypedObjectFromEntries.ts
@@ -1,8 +1,8 @@
 /**
- *
- * @description
+ * 
+ * @description 
  * Object.fromEntries(
- KEYS.map((day, index) => {
+      KEYS.map((day, index) => {
         return [day, VALUES[index]];
       })
  * ) → 이 메서드를 사용할 때 key 타입 추론이 되지 않는 문제를 해결하기 위해 구현
@@ -11,19 +11,19 @@
  * @example getTypedObjectFromEntries(KEYS, VALUES)
  * @returns {Object} An object with key-value pairs.
  * @property key: value | value | value | ...
- * @property key: value | value | value | ...
+ * @property key: value | value | value | ... 
  * @property key: value | value | value | ...
  * @example getTypedObjectFromEntries(
- KEYS,
- KEYS.map(() => OBJECTS)
- )
+      KEYS,
+      KEYS.map(() => OBJECTS)
+    )
  * @returns {Object} An object with key-value pairs.
  * @property key: value[OBJECT, {...}, ...] (= key: object[])
- * @property key: value[OBJECT, {...}, ...]
+ * @property key: value[OBJECT, {...}, ...] 
  * @property key: value[OBJECT, {...}, ...]
  */
 
-export const getTypedObjectFromEntries = <T extends string | symbol | number, K>(
+export const getTypedObjectFromEntries = <T extends string | symbol, K>(
   keys: readonly T[],
   values: readonly K[]
 ): { [Key in T]: K } =>
